chore(documentation): Rename project to Startpunkt
