+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,6 +1,6 @@
 ---
 blank_issues_enabled: false
 contact_links:
-  - name: Startpage Community Support
-    url: https://github.com/ullbergm/startpage/discussions
+  - name: Startpunkt Community Support
+    url: https://github.com/ullbergm/startpunkt/discussions
     about: Please ask and answer questions here.
