+++ src/main/docker/Dockerfile.legacy-jar
@@ -7,11 +7,11 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.legacy-jar -t quarkus/startpage-legacy-jar .
+# docker build -f src/main/docker/Dockerfile.legacy-jar -t quarkus/startpunkt-legacy-jar .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/startpage-legacy-jar
+# docker run -i --rm -p 8080:8080 quarkus/startpunkt-legacy-jar
 #
 # If you want to include the debug port into your docker image
 # you will have to expose the debug port (default 5005 being the default) like this :  EXPOSE 8080 5005.
@@ -20,7 +20,7 @@
 #
 # Then run the container using :
 #
-# docker run -i --rm -p 8080:8080 quarkus/startpage-legacy-jar
+# docker run -i --rm -p 8080:8080 quarkus/startpunkt-legacy-jar
 #
 # This image uses the `run-java.sh` script to run the application.
 # This scripts computes the command line to execute your Java application, and
