+++ src/main/docker/Dockerfile.native-micro
@@ -10,11 +10,11 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.native-micro -t quarkus/startpage .
+# docker build -f src/main/docker/Dockerfile.native-micro -t ullbergm/startpunkt .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/startpage
+# docker run -i --rm -p 8080:8080 quarkus/startpunkt
 #
 ###
 FROM quay.io/quarkus/quarkus-micro-image:2.0
