+++ src/main/java/us/ullberg/startpunkt/health/KubernetesHealthCheck.java
@@ -1,4 +1,4 @@
-package us.ullberg.startpage.health;
+package us.ullberg.startpunkt.health;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientBuilder;
