+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 # OpenShift configuration
-startpage.openshift.enabled=false
+startpunkt.openshift.enabled=false
 
 # Hajimari configuration
-startpage.hajimari.enabled=true
+startpunkt.hajimari.enabled=true
 
 quarkus.micrometer.export.prometheus.enabled=true
 quarkus.micrometer.export.prometheus.path=metrics
@@ -12,4 +12,4 @@ quarkus.cache.caffeine.expire-after-write=30s
 quarkus.quinoa.package-manager-install.node-version=20.11.1
 quarkus.quinoa.package-manager-install.npm-version=10.8.1
 quarkus.quinoa.dev-server.port=3000
-quarkus.quinoa.build-dir=dist
+quarkus.quinoa.build-dir=build
